+++ NEWS.md
@@ -1,5 +1,9 @@
 # store 0.2.0
 
+* Added **get_mysql_connection** and **get_sqlserver_connection** to connect to 
+these databases
+* Added functions to use the **keyring** package to store securely store database 
+credentials
 * Added **extract_comma_delimited_list** and **extract_semicolon_delimited_list** functions
 
 # store 0.1.0